added part 2 to the solution

+++ src/Boat.java
@@ -1,10 +1,10 @@
 public class Boat {
 
-    public int calculatePossibleWays(int duration, int recordDistance) {
-        int possibilities = 0;
-        for (int holdButton = 0; holdButton < duration; holdButton++) {
-            int remainingTime = duration - holdButton - 1;
-            int distance = (holdButton + 1) * remainingTime;
+    public long calculatePossibleWays(long duration, long recordDistance) {
+        long possibilities = 0;
+        for (long holdButton = 0; holdButton < duration; holdButton++) {
+            long remainingTime = duration - holdButton - 1;
+            long distance = (holdButton + 1) * remainingTime;
 
             if (distance > recordDistance) {
                 possibilities++;
